// src/app/images.enum.ts
export enum Images {
  Cloudy = '../assets/svg/weather/021-cloudy-1.svg',
  CloudyWindy = '../assets/svg/weather/021-cloud.svg',
  PartlyCloudyWindy = '../assets/svg/weather/021-cloudy.svg',
  Clear = '../assets/svg/weather/021-sun.svg',
  Rain = '../assets/svg/weather/021-rain.svg',
  RainHighClouds = '../assets/svg/weather/021-rain-2.svg',
  RainLowClouds = '../assets/svg/weather/021-rain-1.svg',
  Sleet = '../assets/svg/weather/021-snowing-1.svg',
  Snowflake = '../assets/svg/weather/021-snowflake.svg',
  Snowing = '../assets/svg/weather/021-snowing-3.svg',
  Storm = '../assets/svg/weather/021-storm.svg',
  Sun = '../assets/svg/weather/021-sun.svg',
  Sunset = '../assets/svg/weather/021-sunset.svg',
  Tornado = '../assets/svg/weather/021-tornado.svg'
}